refactor(ai): isolate reasoning options
